fix(avalonia): address Copilot PR review on #1600 (CallTalk prompt, full-exception logging, alloc-size cast)
- ParamLabel_Click: select the alloc-confirmation prompt by kind — CallTalk
  uses WF "新規にデータを作成しますか？" (AIASMCALLTALKForm.AllocIfNeed) while
  Coordinate/Range keep "新規に座標データを作成しますか？" (both keys already
  translated en/zh/ja, no new translate entry).
- Log the full exception ({inner}/{ex}, i.e. ToString) instead of only
  .Message so jump failures are diagnosable.
- AIScriptPointerJumpCore: cast ASM_ALLOC_SIZE to int at the byte[] alloc
  for clarity/portability.

+++ FEBuilderGBA.Core/AIScriptPointerJumpCore.cs
@@ -161,7 +161,7 @@ namespace FEBuilderGBA
         /// </summary>
         static uint AppendZeroBlock(ROM rom, Undo.UndoData? undodata)
         {
-            byte[] data = new byte[ASM_ALLOC_SIZE]; // all zero
+            byte[] data = new byte[(int)ASM_ALLOC_SIZE]; // all zero
 
             uint newaddr;
             if (CoreState.AppendBinaryData != null && undodata != null)
